Added aircraft image api
