+++ routes/appRoutes.js
@@ -1,6 +1,7 @@
 import { SignUpUser } from "../controllers/users/UsersController"
 import { SearchFlights, AddFlightToUserAccount, AddNewFlight } from "../controllers/flights/FlightsControllers"
 import { AddAirline } from "../controllers/airlines/AirlinesControllers"
+import { GetAircraftImage } from "../controllers/aircrafts/AircraftControllers"
 
 export const routes = (app) => {
   // Public
@@ -15,6 +16,9 @@ export const routes = (app) => {
 
   app.route("/flight/add")
   .post(AddFlightToUserAccount)
+
+  app.route("/aircraft/images")
+  .get(GetAircraftImage)
 
   app.route("/flights")
   .post(AddNewFlight)
